+++ src/main/java/org/greedy/ddarahang/db/country/Country.java
@@ -9,7 +9,6 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import lombok.Getter;
-import org.greedy.ddarahang.db.locationType.LocationType;
 
 @Entity
 @Getter
